+++ lib/td_cx/events.ex
@@ -3,7 +3,10 @@ defmodule TdCx.Events do
   The Events context.
   """
 
+  alias Ecto.Multi
+  alias TdCx.Auth.Claims
   alias TdCx.Events.Event
+  alias TdCx.Jobs.Audit
   alias TdCx.Search.IndexWorker
   alias TdDd.Repo
 
@@ -41,24 +44,33 @@ defmodule TdCx.Events do
 
   ## Examples
 
-      iex> create_event(%{field: value})
+      iex> create_event(%{field: value}, %Claims{})
       {:ok, %Event{}}
 
-      iex> create_event(%{field: bad_value})
+      iex> create_event(%{field: bad_value}, %Claims{})
       {:error, %Ecto.Changeset{}}
 
   """
-  def create_event(attrs \\ %{}) do
-    %Event{}
-    |> Event.changeset(attrs)
-    |> Repo.insert()
+  def create_event(attrs, %Claims{user_id: user_id}) do
+    changeset = Event.changeset(%Event{}, attrs)
+
+    Multi.new()
+    |> Multi.insert(:event, changeset)
+    |> Multi.run(:source_id, fn _, %{event: event} -> {:ok, get_source_id(event)} end)
+    |> Multi.run(:audit, Audit, :job_status_updated, [user_id])
+    |> Repo.transaction()
     |> case do
-      {:ok, %Event{} = event} ->
+      {:ok, %{event: %Event{} = event}} ->
         IndexWorker.reindex(event.job_id)
         {:ok, event}
 
-      error ->
-        error
+      {:error, _, changeset, _} ->
+        {:error, changeset}
     end
+  end
+
+  defp get_source_id(event) do
+    %{job: %{source_id: source_id}} = Repo.preload(event, :job)
+    source_id
   end
 end